Chatbot/requirements.txt
langchain==0.3.3
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-openai==0.2.2
sentence-transformers==3.1.1
python-dotenv==1.0.1
pdfplumber==0.11.4
gradio==5.4.0
